requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==4.0.4
djangorestframework==3.13.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
